move manager termination tests to `test_integration_termination`
A simplified them a bit while at it.
